+++ packages/vinext/src/shims/font-google-base.ts
@@ -1,5 +1,12 @@
 import { buildGoogleFontsUrl as buildUrlFromAxes } from "../build/google-fonts/build-url.js";
-import { formatFontClassRule, resolveSingleFaceStyle, type FontStyle } from "./font-utils.js";
+import {
+  escapeCSSString,
+  formatFontClassRule,
+  resolveSingleFaceStyle,
+  sanitizeCSSVarName,
+  sanitizeFallback,
+  type FontStyle,
+} from "./font-utils.js";
 
 /**
  * next/font/google shim
@@ -19,64 +26,6 @@ import { formatFontClassRule, resolveSingleFaceStyle, type FontStyle } from "./f
  *   // inter.style -> { fontFamily: "'Inter', 'Inter Fallback'", fontStyle: "normal" }
  *   // inter.variable -> CSS class that sets the font CSS variable when requested
  */
-
-/**
- * Escape a string for safe interpolation inside a CSS single-quoted string.
- *
- * Prevents CSS injection by escaping characters that could break out of
- * a `'...'` CSS string context: backslashes, single quotes, and newlines.
- */
-function escapeCSSString(value: string): string {
-  return value
-    .replace(/\\/g, "\\\\")
-    .replace(/'/g, "\\'")
-    .replace(/\n/g, "\\a ")
-    .replace(/\r/g, "\\d ");
-}
-
-/**
- * Validate a CSS custom property name (e.g. `--font-inter`).
- *
- * Custom properties must start with `--` and only contain alphanumeric
- * characters, hyphens, and underscores. Anything else could be used to
- * break out of the CSS declaration and inject arbitrary rules.
- *
- * Returns the name if valid, undefined otherwise.
- */
-function sanitizeCSSVarName(name: string): string | undefined {
-  if (/^--[a-zA-Z0-9_-]+$/.test(name)) return name;
-  return undefined;
-}
-
-/**
- * Sanitize a CSS font-family fallback name.
- *
- * Generic family names (sans-serif, serif, monospace, etc.) are used as-is.
- * Named families are wrapped in escaped quotes. This prevents injection via
- * crafted fallback values like `); } body { color: red; } .x {`.
- */
-function sanitizeFallback(name: string): string {
-  // CSS generic font families — safe to use unquoted
-  const generics = new Set([
-    "serif",
-    "sans-serif",
-    "monospace",
-    "cursive",
-    "fantasy",
-    "system-ui",
-    "ui-serif",
-    "ui-sans-serif",
-    "ui-monospace",
-    "ui-rounded",
-    "emoji",
-    "math",
-    "fangsong",
-  ]);
-  const trimmed = name.trim();
-  if (generics.has(trimmed)) return trimmed;
-  // Wrap in single quotes with escaping to prevent CSS injection
-  return `'${escapeCSSString(trimmed)}'`;
-}
 
 // Track which font stylesheets have been injected (SSR + client)
 const injectedFonts = new Set<string>();
@@ -460,6 +409,9 @@ export function createFontLoader(family: string): FontLoader {
   return function fontLoader(options: FontLoaderOptions = {}): FontResult {
     const internal = options._vinext?.font;
     const fallback = options.fallback ?? [];
+    // The adjusted fallback family name must match the font-family emitted by
+    // buildFallbackFontFace() in build/google-fonts/fallback-metrics.ts ('{family} Fallback').
+    // Keep these two sites in sync to prevent silent fallback mismatches.
     const adjustedFallback =
       options.adjustFontFallback === false || !internal?.adjustedFallbackCSS
         ? []
