+++ packages/vinext/src/shims/font-local.ts
@@ -16,69 +16,14 @@
  *   // myFont.variable -> generated class name when requested
  */
 import {
+  escapeCSSString,
   formatFontClassRule,
   resolveSingleFaceStyle,
+  sanitizeCSSVarName,
+  sanitizeFallback,
   sanitizeFontDescriptorValue,
   type FontStyle,
 } from "./font-utils.js";
-
-/**
- * Escape a string for safe interpolation inside a CSS single-quoted string.
- *
- * Prevents CSS injection by escaping characters that could break out of
- * a `'...'` CSS string context: backslashes, single quotes, and newlines.
- */
-function escapeCSSString(value: string): string {
-  return value
-    .replace(/\\/g, "\\\\")
-    .replace(/'/g, "\\'")
-    .replace(/\n/g, "\\a ")
-    .replace(/\r/g, "\\d ");
-}
-
-/**
- * Validate a CSS custom property name (e.g. `--font-inter`).
- *
- * Custom properties must start with `--` and only contain alphanumeric
- * characters, hyphens, and underscores. Anything else could be used to
- * break out of the CSS declaration and inject arbitrary rules.
- *
- * Returns the name if valid, undefined otherwise.
- */
-function sanitizeCSSVarName(name: string): string | undefined {
-  if (/^--[a-zA-Z0-9_-]+$/.test(name)) return name;
-  return undefined;
-}
-
-/**
- * Sanitize a CSS font-family fallback name.
- *
- * Generic family names (sans-serif, serif, monospace, etc.) are used as-is.
- * Named families are wrapped in escaped quotes. This prevents injection via
- * crafted fallback values like `); } body { color: red; } .x {`.
- */
-function sanitizeFallback(name: string): string {
-  // CSS generic font families — safe to use unquoted
-  const generics = new Set([
-    "serif",
-    "sans-serif",
-    "monospace",
-    "cursive",
-    "fantasy",
-    "system-ui",
-    "ui-serif",
-    "ui-sans-serif",
-    "ui-monospace",
-    "ui-rounded",
-    "emoji",
-    "math",
-    "fangsong",
-  ]);
-  const trimmed = name.trim();
-  if (generics.has(trimmed)) return trimmed;
-  // Wrap in single quotes with escaping to prevent CSS injection
-  return `'${escapeCSSString(trimmed)}'`;
-}
 
 /**
  * Validate a CSS property name for use in declarations.
@@ -90,16 +35,6 @@ function sanitizeFallback(name: string): string {
 function sanitizeCSSProperty(prop: string): string | undefined {
   if (/^(--)?[a-zA-Z][a-zA-Z0-9-]*$/.test(prop)) return prop;
   return undefined;
-}
-
-/**
- * Sanitize a CSS property value for use in declarations.
- *
- * Rejects values containing characters that could break out of a CSS
- * declaration: `{`, `}`, `;`, and `</` (to prevent closing style tags).
- */
-function sanitizeCSSValue(value: string): string | undefined {
-  return sanitizeFontDescriptorValue(value);
 }
 
 let classCounter = 0;
@@ -138,8 +73,8 @@ function generateFontFaceCSS(
   const rules: string[] = [];
 
   for (const src of sources) {
-    const weight = sanitizeCSSValue(src.weight ?? options.weight ?? "400") ?? "400";
-    const style = sanitizeCSSValue(src.style ?? options.style ?? "normal") ?? "normal";
+    const weight = sanitizeFontDescriptorValue(src.weight ?? options.weight ?? "400") ?? "400";
+    const style = sanitizeFontDescriptorValue(src.style ?? options.style ?? "normal") ?? "normal";
     const format = src.path.endsWith(".woff2")
       ? "woff2"
       : src.path.endsWith(".woff")
@@ -163,7 +98,7 @@ function generateFontFaceCSS(
   if (options.declarations) {
     for (const decl of options.declarations) {
       const safeProp = sanitizeCSSProperty(decl.prop);
-      const safeValue = sanitizeCSSValue(decl.value);
+      const safeValue = sanitizeFontDescriptorValue(decl.value);
       if (safeProp && safeValue) {
         rules.push(
           `@font-face { font-family: '${escapeCSSString(family)}'; ${safeProp}: ${safeValue}; }`,
